Rearrange project folders
