+++ Source/Internal/MagneticNotchOverlayBehaviorValue.swift
